gpu: move kernel-scope construction into its own unit
Deciding which symbols the kernel is handed and under what names, giving
an ASSOCIATE or BLOCK scope inside the loop a scope of its own to travel
in, copying a statement into the kernel, hoisting a struct-element gather
onto the host, and pointing every reference in the copies at the kernel's
own symbols, are one job: building the kernel's scope out of the loop's.

What is left in gpu_offload.cpp is the decision itself and the launch:
visit_OMPRegion, the decline reporting it uses, and the pass entry point.
Include lists are pruned to what each file names.

Pure code motion: the definitions are unchanged, only their home is.
